OSGetThreadPriority takes an OSThread*, and two banked files stop faking it
Found by the batch authoring dNandThread_c::create, which could not compile
against the old declaration: the target has bl OSGetCurrentThread immediately
followed by bl OSGetThreadPriority with the result still live in r3.

The header declared `int OSGetThreadPriority(void)`, and d_main.cpp and
d_system.cpp were reproducing the retail sequence by calling OSGetCurrentThread
on its own line and relying on r3 surviving into a zero-argument call. That is
a workaround for a wrong declaration, not a decompilation. Both call sites now
pass the thread they mean, and the emitted code is unchanged - all five
binaries verify.

Also: EGG::Thread's constructor takes `unsigned long`, not u32. The symbol is
__ct__Q23EGG6ThreadFUliiPQ23EGG4Heap and Ul is unsigned long, while u32 is
`unsigned int` here and would mangle Ui. Provable by name, unlike the six
return types this unit has already corrected.

And mMutex's constructor calls OSInitMutex/OSInitCond in its own body: hoisting
them into the caller puts the two vtable stores back to back and MWCC
eliminates the first as dead, which is not what the target's compile did.

+++ include/lib/egg/core/eggThread.h
@@ -8,7 +8,9 @@ class Heap;
 
 class Thread {
 public:
-    Thread(u32 stackSize, int msgCount, int priority, Heap *heap);
+    /// @note `unsigned long`, not `u32`. u32 is `unsigned int` here and would
+/// mangle Ui; the symbol is __ct__Q23EGG6ThreadFUliiPQ23EGG4Heap.
+    Thread(unsigned long stackSize, int msgCount, int priority, Heap *heap);
     Thread(OSThread *, int);
 
     /// @note run(), onEnter() and onExit() MUST keep their inline bodies. The
